+++ flying-saucer-pdf-itext5/src/main/java/org/xhtmlrenderer/pdf/ITextFontResolver.java
@@ -315,26 +315,14 @@ public class ITextFontResolver implements FontResolver {
         File f = new File(path);
         if (f.exists()) {
             ByteArrayOutputStream result = new ByteArrayOutputStream((int)f.length());
-            InputStream is = null;
-            try {
-                is = newInputStream(Paths.get(path));
+            
+            try (InputStream is = newInputStream(Paths.get(path))) {
                 byte[] buf = new byte[10240];
                 int i;
                 while ( (i = is.read(buf)) != -1) {
                     result.write(buf, 0, i);
                 }
-                is.close();
-                is = null;
-
                 return result.toByteArray();
-            } finally {
-                if (is != null) {
-                    try {
-                        is.close();
-                    } catch (IOException e) {
-                        // ignore
-                    }
-                }
             }
         } else {
             throw new IOException("File " + path + " does not exist or is not accessible");
