+++ flying-saucer-pdf/pom.xml
@@ -21,6 +21,7 @@
     <junit.version>4.13.1</junit.version>
     <openpdf.version>1.3.30</openpdf.version>
     <itext.version>2.1.7</itext.version>
+    <slf4j.version>2.0.9</slf4j.version>
   </properties>
 
   <licenses>
@@ -35,6 +36,11 @@
       <groupId>org.xhtmlrenderer</groupId>
       <artifactId>flying-saucer-core</artifactId>
       <version>${project.version}</version>
+    </dependency>
+    <dependency>
+      <groupId>org.slf4j</groupId>
+      <artifactId>slf4j-api</artifactId>
+      <version>${slf4j.version}</version>
     </dependency>
     <dependency>
       <groupId>junit</groupId>
