+++ flying-saucer-pdf/src/main/java/org/xhtmlrenderer/pdf/ITextFSFont.java
@@ -19,12 +19,11 @@
  */
 package org.xhtmlrenderer.pdf;
 
-import org.xhtmlrenderer.pdf.ITextFontResolver.FontDescription;
 import org.xhtmlrenderer.render.FSFont;
 
 public class ITextFSFont implements FSFont {
-    private FontDescription _font;
-    private float _size;
+    private final FontDescription _font;
+    private final float _size;
 
     public ITextFSFont(FontDescription font, float size) {
         _font = font;
